backend/requirements.txt
fastapi
pydantic
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
gunicorn